requirements.txt
pytest==8.4.0
openai==1.86.0
google-auth==2.40.2
google-genai==1.16.1